Shiny application for viewing Atlantis forcings data

A shiny application that displays water exchanges, salinity and temperature data from the netCDF files used as forcings for an Atlantis run.

To run this application a list object must first be generated using make.sh.forcings.object (see Examples).

The Connections tab displays the number of neighbouring water layers that each water layer in a chosen box exchanges water. Use this tab to check that all water layers in boxes exchange with at least one other water layer.

The Exchanges tab allows the user to display the exchanges read into Atlantis (Raw) and all exchanges associated with a box (All). Atlantis only reads in a single exchange value associated with a pair of box layers, however this exchange influences flow for both box layers. Use the (All) tab to see all flows associated with a box. An option is available to plot either raw exchange values or their sign, and whether box numbering is displayed. Exchanges are also presented in a table.

The Horizontal flows tab provides visualisation of the horizontal flow field at each water layer. This flow is approximate as it is calculated using only exchanges at a single layer, however some horizontal flow may occur from horizontal exchanges at the other layers. The length of the arrow only provides a guide to flow speed.

The Vertical flows tab provides net vertical fluxes for each water layer in a box. An option is available to plot either raw exchange values or their sign, and whether box numbering is displayed. Exchanges (per unit area per time step) are also presented in a table.

The Time-series tab allows the user to examine all water exchanges from a specified water layer. As vertical exchanges are often much greater than horizontal exchanges an option is available to suppress plotting of vertical exchanges. This plot can take some time to generate so a checkbox is provided to enable the plot.

The Temperature and Salinity tabs provide spatial and time-series plots. Tabular versions of the data are also presented.

sh.forcings(input.object)

Arguments

input.object

An R list object generated from make.sh.forcings.object.

Value

An object of class 'shiny.appobj' see shinyApp.

Examples

if (FALSE) {
exchange.file    <- "GBR108_hydro.nc"
salinity.file    <- "GBR108_salt.nc"
temperature.file <- "GBR108_temp.nc"
bgm.file         <- "gbr_box_03012012.bgm"
cum.depth <- c(0,5,10,20,50,100,200,3000) # cumulative water layer depths

input.object <- make.sh.forcings.object(
  bgm.file         = bgm.file,
  exchange.file    = exchange.file,
  cum.depth        = cum.depth,
  temperature.file = temperature.file,
  salinity.file    = salinity.file
)
sh.forcings(input.object)
}