Shiny application for viewing Atlantis biological parameters

Uses data from an Atlantis box geometry model file, a group file, and a biological parameter file and provides a visualisation of the data in the form of a shiny application. The three data files must first be pre-processed by make.sh.prm.object, which generates a list object that is the parameter to sh.prm (see Examples).

The Map tab displays the spatial distribution of box depths.

The Groups tab displays the information stored in the Atlantis group file.

The Global params tab shows the global variables that have values provided in the biological parameters file.

The Group params tab allows the user to see a list of possible parameters associated with groups (Definitions) and the values provided in the biological parameters file (Values). Use the checkboxes to view the group parameters of interest.

The Habitat tab shows which groups are associated with habitats.

The Distributions tab shows horizontal and vertical distributions for chosen groups. Initial values, recruitment and migration parameters can be viewed here.

The Feeding tab shows clearance rates, maximum growth rates, prey availability and refuge-related parameters.

The Migration tab provides migrant biomasses (structural and reserve) and numbers.

The Recruitment tab shows the relation between structural and reserve biomass across all groups.

sh.prm(obj)

Arguments

obj

R list object generated from make.sh.prm.object

Value

object of class 'shiny.appobj' see shinyApp

Examples

if (FALSE) {
bgm.file <- "SEAP_extended_shelf.bgm"
grp.file <- "SEAP_Groups_Aquacult.csv"
prm.file <- "SEAP_biol_pH_Aquacult.prm"
obj <- make.sh.prm.object(bgm.file, grp.file, prm.file)
sh.prm(obj)
}